style: 즐겨찾기 기능 리팩토링을 위한 코드 수정

+++ src/main/java/com/example/ajouparking/service/LikesService.java
@@ -16,7 +16,7 @@ public class LikesService {
     private final UserRepository userRepository;
 
     @Transactional
-    public void putLike(int fromUserId, Long toReviewId){
+    public void putLike(long fromUserId, long toReviewId){
         try{
             User user = userRepository.findById(fromUserId)
                     .orElseThrow(() -> new IllegalArgumentException("없는유저"));
@@ -31,7 +31,7 @@ public class LikesService {
     }
 
     @Transactional
-    public void deleteLike(int fromUserId, Long toUserId){
+    public void deleteLike(long fromUserId, long toUserId){
         User user = userRepository.findById(fromUserId)
                 .orElseThrow(() -> new IllegalArgumentException("없는유저"));
